Move the purging methods for EventStream into a module.
