+++ app/models/event_stream.rb
@@ -1,4 +1,5 @@
 class EventStream < ApplicationRecord
+  include_concern 'Purging'
   serialize :full_data
 
   belongs_to :target, :polymorphic => true
@@ -31,49 +32,5 @@ class EventStream < ApplicationRecord
     Notification.emit_for_event(self)
   rescue => err
     _log.log_backtrace(err)
-  end
-
-  #
-  # Purging methods
-  #
-
-  def self.keep_ems_events
-    ::Settings.ems_events.history.keep_ems_events
-  end
-
-  def self.purge_date
-    keep = keep_ems_events.to_i_with_method.seconds
-    keep = 6.months if keep == 0
-    keep.ago.utc
-  end
-
-  def self.purge_window_size
-    ::Settings.ems_events.history.purge_window_size
-  end
-
-  def self.purge_timer
-    purge_queue(purge_date)
-  end
-
-  def self.purge_queue(ts)
-    MiqQueue.put(
-      :class_name  => name,
-      :method_name => "purge",
-      :role        => "event",
-      :queue_name  => "ems",
-      :args        => [ts],
-    )
-  end
-
-  def self.purge(older_than, window = nil, limit = nil)
-    _log.info("Purging #{limit || "all"} events older than [#{older_than}]...")
-
-    window ||= purge_window_size
-
-    total = where(arel_table[:timestamp].lteq(older_than)).delete_in_batches(window, limit) do |count, _total|
-      _log.info("Purging #{count} events.")
-    end
-
-    _log.info("Purging #{limit || "all"} events older than [#{older_than}]...Complete - Deleted #{total} records")
   end
 end
